Move error bound tests to separate file
